Only allow missing values for subscale scores.
We are currently workign with AIHW to redefine how the YES-PHN subscales
are calculated.

For our measure specs, we allow the subscales to be "missing" and we
calculate them on upload / data entry. Forcing "missing" subscale values
is consistent here; allowing us to then recalculate all yes-pgn
subscales once the work with AIHW is complete.